chore: add payment message component and fix minor mistakes
