--- packages/ui-components/src/pages/Checkout/index.ts
@@ -1,1 +1,0 @@
-export { default as Checkout } from './Checkout';
