--- packages/ui-components/src/pages/PDP/index.ts
@@ -1,1 +1,0 @@
-export { default as PDP } from './PDP';
