+++ packages/ui-components/src/pages/PaymentMessage/index.tsx
@@ -1,0 +1,1 @@
+export { default as PaymentMessage } from './PaymentMessage';
