+++ packages/ui-components/src/pages/index.ts
@@ -1,5 +1,3 @@
 export * from './PLP';
-export * from './PDP';
 export * from './CLP';
-export * from './Checkout';
-export * from './Cart';
+export * from './PaymentMessage';
